Added first functional route for datasets
