+++ DashAI/back/Dataloaders/dataloaders_schemas/splits_settings.json
@@ -56,5 +56,5 @@
 			}
 		]
 	},
-	
+
 }
