+++ DashAI/back/Dataloaders/dataloaders_schemas/tabular_datasets.json
@@ -64,5 +64,5 @@
 			}
 		]
 	},
-	
+
 }
